+++ src/app/rounds/page.tsx
@@ -1,0 +1,5 @@
+import { Rounds } from "@/components/home/Rounds"
+
+export default function Page() {
+  return <Rounds />
+}
