+++ src/components/common/Navbar.tsx
@@ -1,5 +1,6 @@
 "use client"
 import Image from "next/image"
+import Link from "next/link"
 import React from "react"
 
 import { Account } from "./Account"
@@ -7,13 +8,15 @@ import { Account } from "./Account"
 const Navbar = () => {
   return (
     <nav className="sticky inset-x-0 top-0 h-18 bg-white flex items-center justify-between px-6 shadow-sm z-20">
-      <Image
-        src="/assets/images/logo.svg"
-        height={24}
-        width={167}
-        priority
-        alt=""
-      />
+      <Link href="/">
+        <Image
+          src="/assets/images/logo.svg"
+          height={24}
+          width={167}
+          priority
+          alt=""
+        />
+      </Link>
       <Account />
     </nav>
   )
