+++ src/components/home/FundingRounds.tsx
@@ -4,6 +4,7 @@ import { format } from "date-fns"
 import { ArrowRight } from "lucide-react"
 import Image from "next/image"
 import { useRouter } from "next/navigation"
+import { useSession } from "next-auth/react"
 import { useMemo } from "react"
 import { optimism } from "viem/chains"
 
@@ -82,11 +83,16 @@ const Round = ({
   link?: string
 }) => {
   const router = useRouter()
+  const { status } = useSession()
   const { setOpenDialog } = useAppDialogs()
 
   const onClick = () => {
     if (fundingRound.status === "now") {
-      setOpenDialog("get_started")
+      if (status === "authenticated") {
+        router.push("/dashboard")
+      } else {
+        setOpenDialog("get_started")
+      }
       return
     }
